+++ personread/src/test/java/software/daveturner/personread/service/PersonReadServiceImplTest.java
@@ -20,18 +20,39 @@ class PersonReadServiceImplTest {
     @MockBean
     PersonReadRepo repo;
 
+    @MockBean
+    RedisService redisService;
+
     @Test
-    public void ensureFetchByIdReturnsExpected() {
+    public void ensureFetchByIdReturnsExpectedWithRedisResponse() {
         Person person = new Person();
         person.setId("123");
 
         Optional<Person> mockedData = Optional.of(person);
         doReturn(mockedData).when(repo).findById(anyString());
+        doReturn(mockedData).when(redisService).getPerson(anyString());
+
+        Optional<Person> value = service.findById("123");
+        assertTrue(value.isPresent());
+        assertEquals(value.get().getId(), person.getId());
+
+        verify(repo, times(0)).findById(anyString());
+        verify(redisService, times(1)).getPerson(anyString());
+    }
+
+    @Test
+    public void ensureFetchByIdReturnsExpectedWithMongoResponse() {
+        Person person = new Person();
+        person.setId("123");
+        Optional<Person> mockedData = Optional.of(person);
+        doReturn(mockedData).when(repo).findById(anyString());
+        doReturn(Optional.empty()).when(redisService).getPerson(anyString());
 
         Optional<Person> value = service.findById("123");
         assertTrue(value.isPresent());
         assertEquals(value.get().getId(), person.getId());
 
         verify(repo, times(1)).findById(anyString());
+        verify(redisService, times(1)).getPerson(anyString());
     }
 }